tryout auth

+++ server/quiz-router.js
@@ -8,6 +8,9 @@ export default function create(directory) {
     const quizService = new QuizService(directory)
 
     router.post('/', async (req, res) => {
+        if (!req.isAuthenticated()) {
+            return res.status(401).end('Not authenticated!')
+        }
         try {
             const author = req.user.id
             const id = await quizService.create(req.body, author)
@@ -20,6 +23,9 @@ export default function create(directory) {
     })
 
     router.get('/', async (req, res) => {
+        if (!req.isAuthenticated()) {
+            return res.status(401).end('Not authenticated!')
+        }
         try {
             const author = req.user.id
             let quizList
@@ -36,6 +42,9 @@ export default function create(directory) {
     })
 
     router.delete('/:id', async (req, res) => {
+        if (!req.isAuthenticated()) {
+            return res.status(401).end('Not authenticated!')
+        }
         try {
             await quizService.delete(req.params.id)
             res.status(200).end()
@@ -46,6 +55,9 @@ export default function create(directory) {
     })
 
     router.get('/:id', async (req, res) => {
+        if (!req.isAuthenticated()) {
+            return res.status(401).end('Not authenticated!')
+        }
         try {
             const quiz = await quizService.get(req.params.id)
             res.status(200).send(quiz).end()
@@ -56,6 +68,9 @@ export default function create(directory) {
     })
 
     router.put('/:id', async (req, res) => {
+        if (!req.isAuthenticated()) {
+            return res.status(401).end('Not authenticated!')
+        }
         try {
             const author = req.user.id
             await quizService.update(req.params.id, req.body, author)
